Gemini/Ask Ask/requirements.txt
flask
flask-cors
google-genai
requests
